+++ applications/fishing-map/src/features/map/popups/VesselEventsLayers.tsx
@@ -4,6 +4,7 @@ import { groupBy } from 'lodash'
 import { DateTime } from 'luxon'
 import { TooltipEventFeature } from 'features/map/map.hooks'
 import { formatI18nDate } from 'features/i18n/i18nDate'
+import { MAX_TOOLTIP_LIST } from '../map.slice'
 import styles from './Popup.module.css'
 
 // t('event.fishing', 'Fished for')
@@ -13,9 +14,11 @@ type ContextTooltipRowProps = {
   features: TooltipEventFeature[]
 }
 
-function EnvironmentTooltipSection({ features }: ContextTooltipRowProps) {
+function VesselEventsTooltipSection({ features }: ContextTooltipRowProps) {
   const { t } = useTranslation()
-  const featuresByType = groupBy(features, 'layerId')
+  const overflows = features?.length > MAX_TOOLTIP_LIST
+  const maxFeatures = overflows ? features.slice(0, MAX_TOOLTIP_LIST) : features
+  const featuresByType = groupBy(maxFeatures, 'layerId')
   return (
     <Fragment>
       {Object.values(featuresByType).map((featureByType, index) => (
@@ -30,7 +33,7 @@ function EnvironmentTooltipSection({ features }: ContextTooltipRowProps) {
                 .diff(DateTime.fromISO(feature.properties.start), ['hours', 'minutes', 'seconds'])
                 .toObject()
               return (
-                <Fragment>
+                <Fragment key={index}>
                   <div className={styles.row}>
                     <span className={styles.rowText}>
                       {formatI18nDate(feature.properties.start, { format: DateTime.DATETIME_FULL })}{' '}
@@ -51,6 +54,12 @@ function EnvironmentTooltipSection({ features }: ContextTooltipRowProps) {
                 </Fragment>
               )
             })}
+
+            {overflows && (
+              <div className={styles.vesselsMore}>
+                + {features.length - MAX_TOOLTIP_LIST} {t('common.more', 'more')}
+              </div>
+            )}
           </div>
         </div>
       ))}
@@ -58,4 +67,4 @@ function EnvironmentTooltipSection({ features }: ContextTooltipRowProps) {
   )
 }
 
-export default EnvironmentTooltipSection
+export default VesselEventsTooltipSection